Fix bug where exception is thrown when MFA is not enabled on Garmin account.
OAuth1 token responses for accounts without MFA do not contain mfa_expiration_timestamp. Previously, when there was no timestamp, the object would be created with a null timestamp. When adding date parsing, invalid values were not considered. A check for null timestamps has been added. If no expiration is given, token is assumed not to expire.